refactor: split model providers by backend
